+++ src/main/java/org/folio/service/FeesFinesEventsGenerationService.java
@@ -1,7 +1,6 @@
 package org.folio.service;
 
 import static io.vertx.core.Future.failedFuture;
-import static io.vertx.core.Future.succeededFuture;
 
 import java.math.BigDecimal;
 import java.util.List;
@@ -9,10 +8,12 @@ import java.util.Map;
 import java.util.stream.Collectors;
 
 import org.folio.repository.SynchronizationJobRepository;
+import org.folio.rest.client.FeesFinesClient;
 import org.folio.rest.handlers.EventHandler;
 import org.folio.rest.handlers.FeeFineBalanceChangedEventHandler;
 import org.folio.rest.jaxrs.model.Account;
 import org.folio.rest.jaxrs.model.FeeFineBalanceChangedEvent;
+import org.folio.rest.jaxrs.model.Feefine;
 import org.folio.rest.jaxrs.model.SynchronizationJob;
 
 import io.vertx.core.Future;
@@ -20,7 +21,9 @@ import io.vertx.core.Vertx;
 import io.vertx.core.json.JsonObject;
 
 public class FeesFinesEventsGenerationService extends EventsGenerationService {
+  public static final String ACCOUNTS = "accounts";
   private final EventHandler<FeeFineBalanceChangedEvent> feeFineBalanceChangedEventHandler;
+  private final FeesFinesClient feeFineOkapiClient;
 
   public FeesFinesEventsGenerationService(Map<String, String> okapiHeaders, Vertx vertx,
     SynchronizationJobRepository syncRepository) {
@@ -28,6 +31,7 @@ public class FeesFinesEventsGenerationService extends EventsGenerationService {
     super(okapiHeaders, vertx, syncRepository);
     this.feeFineBalanceChangedEventHandler = new FeeFineBalanceChangedEventHandler(
       okapiHeaders, vertx);
+    this.feeFineOkapiClient = new FeesFinesClient(vertx, okapiHeaders);
   }
 
   @Override
@@ -36,40 +40,39 @@ public class FeesFinesEventsGenerationService extends EventsGenerationService {
 
     Future<JsonObject> readPage = okapiClient.getMany(path, PAGE_LIMIT, pageNumber * PAGE_LIMIT)
       .compose(jsonPage -> {
-        if (jsonPage == null || jsonPage.getJsonArray("accounts").size() == 0) {
+        if (jsonPage == null || jsonPage.getJsonArray(ACCOUNTS).size() == 0) {
           String errorMessage = String.format(
             "Error in receiving page number %d of accounts: %s", pageNumber, path);
           log.error(errorMessage);
           return failedFuture(errorMessage);
         }
         return generateEventsByAccounts(mapJsonToAccounts(jsonPage))
-          .onComplete(result -> {
-            if (result.succeeded()) {
-              log.info("Success adding to generatedEventsForEachPagesToList for loans");
-              updateSyncJobWithProcessedAccounts(syncJob,
-                syncJob.getNumberOfProcessedFeesFines() + jsonPage.getJsonArray("accounts").size(),
-                totalRecords);
-            } else {
-              log.error("Failure adding to generatedEventsForEachPagesToList for loans");
-              updateSyncJobWithError(syncJob, result.cause().getLocalizedMessage());
-            }
-          })
+          .onComplete(r -> logEventsGenerationResult(r, ACCOUNTS))
+          .compose(r -> updateSyncJobWithProcessedAccounts(syncJob,
+            syncJob.getNumberOfProcessedFeesFines() + jsonPage.getJsonArray(ACCOUNTS).size(),
+            totalRecords))
+          .recover(t -> updateSyncJobWithError(syncJob, t.getLocalizedMessage()))
           .map(jsonPage);
       });
     generatedEventsForPages.add(readPage);
   }
 
   private Future<String> generateEventsByAccounts(List<Account> records) {
-    return records.stream()
-      .map(this::generateFeeFineBalanceChangedEvent)
-      .reduce(Future.succeededFuture(), (a, b) -> a.compose(r -> b));
+    return feeFineOkapiClient.fetchFeeFineTypes()
+      .map(feeFineTypes -> feeFineTypes.stream()
+        .collect(Collectors.toMap(Feefine::getFeeFineType, Feefine::getId)))
+      .compose(feeFineTypes -> records.stream()
+        .map(account -> generateFeeFineBalanceChangedEvent(account,
+          feeFineTypes.get(account.getFeeFineType())))
+        .reduce(Future.succeededFuture(), (a, b) -> a.compose(r -> b)));
   }
 
-  private Future<String> generateFeeFineBalanceChangedEvent(Account account) {
+  private Future<String> generateFeeFineBalanceChangedEvent(Account account, String feeFineTypeId) {
     log.info("Start generateFeeFineBalanceChangedEvent for account " + account.getId());
     return feeFineBalanceChangedEventHandler.handle(new FeeFineBalanceChangedEvent()
       .withBalance(BigDecimal.valueOf(account.getRemaining()))
-      .withFeeFineId(account.getFeeFineId())
+      .withFeeFineId(account.getId())
+      .withFeeFineTypeId(feeFineTypeId)
       .withUserId(account.getUserId())
       .withLoanId(account.getLoanId())
       .withMetadata(account.getMetadata()), true)
@@ -78,7 +81,7 @@ public class FeesFinesEventsGenerationService extends EventsGenerationService {
     }
 
   private List<Account> mapJsonToAccounts(JsonObject loansJson) {
-    return loansJson.getJsonArray("accounts").stream()
+    return loansJson.getJsonArray(ACCOUNTS).stream()
       .filter(obj -> obj instanceof JsonObject)
       .map(JsonObject.class::cast)
       .map(this::mapToAccount)
@@ -96,19 +99,12 @@ public class FeesFinesEventsGenerationService extends EventsGenerationService {
       .withMetadata(mapMetadataFromJson(representation.getJsonObject("metadata")));
   }
 
-  private void updateSyncJobWithProcessedAccounts(SynchronizationJob syncJob, int processed,
-    int total) {
+  private Future<SynchronizationJob> updateSyncJobWithProcessedAccounts(SynchronizationJob syncJob,
+    int processed, int total) {
 
-    SynchronizationJob updatedSyncJob = syncJob
-      .withNumberOfProcessedFeesFines(processed)
+    syncJob.withNumberOfProcessedFeesFines(processed)
       .withTotalNumberOfFeesFines(total);
-    syncRepository.update(updatedSyncJob, syncJob.getId())
-    .onComplete(r -> {
-      if (r.failed()) {
-        log.error("updateSyncJobWithProcessedAccounts failed");
-      } else {
-        log.info("updateSyncJobWithProcessedAccounts success");
-      }
-    });
+
+    return syncRepository.update(syncJob);
   }
 }